+++ price-adapter/src/sources.rs
@@ -1,7 +1,9 @@
 mod band_stable_coin;
 mod binance;
 mod coingecko;
+mod coinmarketcap;
 
 pub use band_stable_coin::*;
 pub use binance::*;
 pub use coingecko::*;
+pub use coinmarketcap::*;