+++ price-adapter/CHANGELOG.md
@@ -3,5 +3,6 @@
 ## Unreleased
 
 - Support CoinGecko sources
+- Support CoinMarketCap source
 - Support Binance Websocket source
 - Create Interval and Websocket service to manage sources